SDIT-2832 Move tests into tap package
